Makes visibility evaluator a model function
